Phase1/day9/requirements.txt
pydantic
python-dotenv
PyPDF2
python-docx